requirements.txt
streamlit
fastapi
uvicorn[standard]
ultralytics
google-auth-oauthlib
google-auth
requests
pillow
python-multipart

